Updated on 2026-08-14

+++ app/src/main/java/com/tangem/data/network/model/InsightResponse.kt
@@ -16,7 +16,7 @@ data class InsightResponse(
         var txid: String = "",
 
         @SerializedName("satoshis")
-        var satoshis: Int? = null,
+        var satoshis: Long? = null,
 
         @SerializedName("height")
         var height: Int? = null,
